Model Refactor: organize models
