+++ app/Filament/Components/FormVersionBuilder.php
@@ -9,8 +9,8 @@ use Filament\Forms\Components\Actions\Action;
 use Filament\Forms\Components\Section;
 use Filament\Forms\Components\Select;
 use Filament\Forms\Components\Hidden;
-use App\Models\StyleSheet;
-use App\Models\FormScript;
+use App\Models\FormBuilding\StyleSheet;
+use App\Models\FormBuilding\FormScript;
 use Filament\Forms\Components\Tabs;
 use Filament\Forms\Components\Tabs\Tab;
 use Filament\Resources\Pages\ViewRecord;
