+++ app/Filament/Forms/Resources/FormResource.php
@@ -24,9 +24,9 @@ use Filament\Infolists\Components\RepeatableEntry;
 use Filament\Infolists\Components\Grid as InfolistGrid;
 use Illuminate\Support\Facades\Gate;
 use App\Models\Ministry;
-use App\Models\FormReach;
-use App\Models\FormFrequency;
-use App\Models\FormRepository;
+use App\Models\FormMetadata\FormReach;
+use App\Models\FormMetadata\FormFrequency;
+use App\Models\FormMetadata\FormRepository;
 use App\Models\UserType;
 use App\Filament\Components\FormDeploymentsManager;
 use Illuminate\Support\Str;
