+++ app/Filament/Forms/Resources/FormVersionResource/Pages/EditFormVersion.php
@@ -3,8 +3,8 @@
 namespace App\Filament\Forms\Resources\FormVersionResource\Pages;
 
 use App\Filament\Forms\Resources\FormVersionResource;
-use App\Models\StyleSheet;
-use App\Models\FormScript;
+use App\Models\FormBuilding\StyleSheet;
+use App\Models\FormBuilding\FormScript;
 use Filament\Actions;
 use Filament\Resources\Pages\EditRecord;
