+++ app/Models/FormMetadata/FormDataSource.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\FormMetadata;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use App\Models\FormBuilding\FormVersion;
 
 class FormDataSource extends Model
 {
@@ -25,5 +26,4 @@ class FormDataSource extends Model
     {
         return $this->belongsToMany(FormVersion::class, 'form_versions_form_data_sources');
     }
-
 }
