+++ app/Models/FormMetadata/FormFrequency.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\FormMetadata;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
+use App\Models\Form;
 
 class FormFrequency extends Model
 {
