+++ app/Models/FormMetadata/FormRepository.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\FormMetadata;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use App\Models\Form;
 
 class FormRepository extends Model
 {
