+++ app/Models/FormMetadata/FormSoftwareSource.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\FormMetadata;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use App\Models\Form;
 
-class FormLocation extends Model
+class FormSoftwareSource extends Model
 {
     use HasFactory;
 
@@ -17,6 +18,6 @@ class FormLocation extends Model
 
     public function forms(): BelongsToMany
     {
-        return $this->belongsToMany(Form::class, 'form_form_location');
+        return $this->belongsToMany(Form::class, 'form_software_source_form');
     }
 }
