+++ app/Models/FormMetadata/FormTag.php
@@ -1,22 +1,23 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\FormMetadata;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use App\Models\Form;
 
-class FormSoftwareSource extends Model
+class FormTag extends Model
 {
     use HasFactory;
-    
+
     protected $fillable = [
         "name",
-        "description"
+        "description",
     ];
 
     public function forms(): BelongsToMany
     {
-        return $this->belongsToMany(Form::class, 'form_software_source_form');
+        return $this->belongsToMany(Form::class, 'form_form_tags');
     }
 }
