+++ app/Providers/AppServiceProvider.php
@@ -7,7 +7,7 @@ use Illuminate\Support\Facades\URL;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Database\Eloquent\Model;
 use App\Models\User;
-use App\Models\FormVersion;
+use App\Models\FormBuilding\FormVersion;
 use App\Observers\FormVersionObserver;
 
 class AppServiceProvider extends ServiceProvider
