+++ database/seeders/FormFrequencySeeder.php
@@ -2,7 +2,7 @@
 
 namespace Database\Seeders;
 
-use App\Models\FormFrequency;
+use App\Models\FormMetadata\FormFrequency;
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
